# perf(libzip): enable more compiler optimizations