… .pnp.cjs
… packages/yarnpkg-core/sources/worker-zip/index.js
… packages/yarnpkg-libzip/sources/libzipAsync.js
… packages/yarnpkg-libzip/sources/libzipSync.js
… packages/yarnpkg-pnp/sources/hook.js