**What's the problem this PR addresses?**

We're building `libzip` without optimizations enabled and `zlib-ng` with optimization level `-O2`.

**How did you fix it?**

Add the `-O3` flag to the build of both `zlib-ng` and `libzip` to let the compiler optimize as much as possible.

**Results**

The performance didn't change much so I'm assuming V8 is able to optimize it well enough on its own but it did reduce the bundle size.

```diff
$ du -b .pnp*
- 3303160 .pnp-before.cjs
+ 3261496 .pnp-after.cjs

$ du -b yarn*
- 2871526 yarn-before.js
+ 2805394 yarn-after.js
```

```console
=== install-full-cold ===
Benchmark 1: node ./before.cjs
  Time (mean ± σ):     41.453 s ±  0.301 s    [User: 82.733 s, System: 4.856 s]
  Range (min … max):   40.997 s … 42.100 s    10 runs

Benchmark 2: node ./after.cjs
  Time (mean ± σ):     41.307 s ±  0.323 s    [User: 82.011 s, System: 4.890 s]
  Range (min … max):   40.789 s … 41.765 s    10 runs

Summary
  'node ./after.cjs' ran
    1.00 ± 0.01 times faster than 'node ./before.cjs'

=== install-cache-only ===
Benchmark 1: node ./before.cjs
  Time (mean ± σ):     22.573 s ±  0.288 s    [User: 23.861 s, System: 1.992 s]
  Range (min … max):   22.037 s … 23.077 s    10 runs

Benchmark 2: node ./after.cjs
  Time (mean ± σ):     22.521 s ±  0.148 s    [User: 23.991 s, System: 1.961 s]
  Range (min … max):   22.253 s … 22.719 s    10 runs

Summary
  'node ./after.cjs' ran
    1.00 ± 0.01 times faster than 'node ./before.cjs'

=== install-cache-and-lock ===
Benchmark 1: node ./before.cjs
  Time (mean ± σ):      5.103 s ±  0.139 s    [User: 6.618 s, System: 0.836 s]
  Range (min … max):    4.882 s …  5.271 s    10 runs

Benchmark 2: node ./after.cjs
  Time (mean ± σ):      5.064 s ±  0.160 s    [User: 6.448 s, System: 0.851 s]
  Range (min … max):    4.640 s …  5.230 s    10 runs

Summary
  'node ./after.cjs' ran
    1.01 ± 0.04 times faster than 'node ./before.cjs'

=== install-ready ===
Benchmark 1: node ./before.cjs add dummy-pkg@link:./dummy-pkg
  Time (mean ± σ):      2.101 s ±  0.024 s    [User: 2.729 s, System: 0.179 s]
  Range (min … max):    2.056 s …  2.135 s    10 runs

Benchmark 2: node ./after.cjs add dummy-pkg@link:./dummy-pkg
  Time (mean ± σ):      2.109 s ±  0.033 s    [User: 2.732 s, System: 0.182 s]
  Range (min … max):    2.082 s …  2.184 s    10 runs

Summary
  'node ./before.cjs add dummy-pkg@link:./dummy-pkg' ran
    1.00 ± 0.02 times faster than 'node ./after.cjs add dummy-pkg@link:./dummy-pkg'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.